requirements.txt
Pillow>=10.2.0
opencv-python>=4.9.0.80
numpy>=1.26.3
pydub>=0.25.1
Flask>=3.0.0
